+++ src/commands/mod/kick.js
@@ -20,22 +20,27 @@ module.exports = class CMD extends Command {
                 disabled: false,
                 cooldown: 10,
                 category: "Moderation",
+                slash: true,
+                options: [
+                    {
+                        name: "user",
+                        description: "Which user to kick",
+                        type: "USER",
+                        required: true
+                    }
+                ]
             },
             client
         );
     }
 
     async execute({ message, args, guildDB }, t) {
-        //TODO: Add translation
         const user = userFromMention(args[0], message.client);
         if (!user) {
             return message.reply(t("errors:invalidUser"));
         }
-        let member = message.guild.members.cache.get(user.id);
-        if (!member) {
-            member = await message.guild.members.fetch(user.id);
+        let member = await message.guild.members.fetch(user.id);
             if (!member) return message.reply(t("errors:userNotInGuild"));
-        }
         if (user.id === message.client.user.id)
             return message.reply(t("cmds:kick.mySelf"));
 
@@ -56,6 +61,7 @@ module.exports = class CMD extends Command {
             member.kick(reason);
         } catch (error) {
             console.error(error);
+            //TODO: add translation
             return message.channel.send(`Failed to kick **${user.tag}**`);
         }
 
@@ -66,7 +72,7 @@ module.exports = class CMD extends Command {
             if (channel) {
                 const embed = new Embed({ color: "red" });
                 embed.setTitle(
-                    `${t("cmds:lick.kicked")}: ${user.tag} (${user.id})`
+                    `${t("cmds:kick.kicked")}: ${user.tag} (${user.id})`
                 );
                 embed.addField(
                     t("misc:resMod"),
@@ -77,7 +83,62 @@ module.exports = class CMD extends Command {
             }
         }
 
+        //TODO: add translation
         return message.channel.send(
+            `Successfully kicked **${user.tag}** from the server!`
+        );
+    }
+
+    async run({ interaction, guildDB }, t) {
+        const user = interaction.options.getUser("user");
+        if (!user) {
+            return interaction.editReply(t("errors:invalidUser"));
+        }
+        const member = await interaction.guild.members.fetch(user.id);
+        if (!member) return interaction.editReply(t("errors:userNotInGuild"));
+        if (user.id === interaction.client.user.id)
+            return interaction.editReply(t("cmds:kick.mySelf"));
+        if (
+            member.roles.highest.position >=
+            interaction.member.roles.highest.position
+        )
+            return interaction.editReply(t("misc:higherRole"));
+
+        if (
+            interaction.guild.me.roles.highest.position <=
+            member.roles.highest.position
+        )
+            return interaction.editReply(t("misc:higherRoleBot"));
+
+        const reason = args.slice(1).join(" ") || t("misc:not_spec");
+        try {
+            member.kick(reason);
+        } catch (error) {
+            console.error(error);
+            //TODO: add translation
+            return interaction.editReply(`Failed to kick **${user.tag}**`);
+        }
+
+        if (guildDB.plugins.modlogs) {
+            const channel = message.guild.channels.cache.get(
+                guildDB.plugins.modlogs
+            );
+            if (channel) {
+                const embed = new Embed({ color: "red" });
+                embed.setTitle(
+                    `${t("cmds:kick.kicked")}: ${user.tag} (${user.id})`
+                );
+                embed.addField(
+                    t("misc:resMod"),
+                    `${message.author.tag} (${message.author.id})`
+                );
+                embed.addField(t("misc:reason"), reason);
+                channel.send({ embeds: [embed] });
+            }
+        }
+
+        //TODO: add translation
+        return interaction.editReply(
             `Successfully kicked **${user.tag}** from the server!`
         );
     }